Fix drag issue
